so this is a tic tac toe game i created. I still need to fix one or two things, such as modifying the optimal strategy of the AI and commenting the code properly.
The end idea would be to have several mini-games that the prof could lunch through a script. I currently am working on tic tac toe and on a small blackjack. I invite you to come up with other minigames, such as snake or sth that can be played in the terminal.